+++ changelog.md
@@ -1,10 +1,15 @@
 Autopilot [handlers] (2020-06-15) 
-  [autopilot_api_v1] 
+  * [autopilot_api_v1] 
     * Initial commit of the handler
+    
 Autopilot [bridge-adapters] (2020-06-15)
-  [kinetic-bridgehub-adapter-autopilot] 
+  * [kinetic-bridgehub-adapter-autopilot] 
     * Initial commit of the adapter
 
 Autopilot [bridge-adapters] (2020-07-17)
-  [kinetic-bridgehub-adapter-autopilot] 
+  * [kinetic-bridgehub-adapter-autopilot] 
     * added JsonPath.
+
+Autopilot [bridge-adapters] (2020-07-20)
+  * [kinetic-bridgehub-adapter-autopilot] 
+    * added a conditional and tests for JsonPath.